updated multiple section

+++ src/modules/vehicles/vehicles.service.ts
@@ -50,13 +50,7 @@ const updateVehicle = async (
   const result = await pool.query(
     `
     UPDATE Vehicles SET
-      vehicle_name = COALESCE($1, vehicle_name),
-      type = COALESCE($2, type),
-      registration_number = COALESCE($3, registration_number),
-      daily_rent_price = COALESCE($4, daily_rent_price),
-      availability_status = COALESCE($5, availability_status)
-    WHERE id = $6
-    RETURNING *
+      vehicle_name = COALESCE($1, vehicle_name), type = COALESCE($2, type), registration_number = COALESCE($3, registration_number), daily_rent_price = COALESCE($4, daily_rent_price), availability_status = COALESCE($5, availability_status) WHERE id = $6 RETURNING *
     `,
     [
       vehicle_name,
@@ -71,6 +65,15 @@ const updateVehicle = async (
 };
 
 const deleteVehicle = async (vehicleId: string) => {
+  const bookingStatus = await pool.query(
+    `
+    SELECT * FROM Bookings WHERE vehicle_id=$1 AND status='active'
+    `,
+    [vehicleId]
+  );
+  if (bookingStatus.rows.length > 0) {
+    throw new Error('Cannot delete vehicle: Vehicle has existing bookings');
+  }
   const result = await pool.query(
     `
     DELETE FROM Vehicles WHERE id=$1
